+++ Samples~/Geodesics/Torus Examples/TorusMovementController.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using UnityEngine;
+using Utilities.Geodesics;
+
+namespace Samples.Geodesics.Sphere
+{
+    public class TorusMovementController : MonoBehaviour
+    {
+
+    }
+}